requirements.txt
streamlit==1.26.0
transformers==4.31.0
langchain==0.0.278
faiss-cpu==1.7.4
PyPDF2==3.0.1
spacy==3.6.1
bert-extractive-summarizer==0.10.1
torch==2.0.1
sentencepiece==0.1.99
